优化 MainViewModel 的代码布局，实现关注点分离
