+++ D5RobotWPF/MainWindow.xaml
@@ -89,7 +89,7 @@
             <TextBlock>点动分辨率</TextBlock>
             <ComboBox SelectedIndex="0" ItemsSource="{Binding JogResolutions, Mode=OneTime}" SelectedItem="{Binding JogResolutionSelected}"/>
             <mah:ToggleSwitch Content="相对位姿" IsOn="{Binding IsPoseRelative}"/>
-            <mah:ToggleSwitch Content="遥操作" IsOn="{Binding EnableGamepad}" Command="{Binding ToggleTeleopCommand}"/>
+            <mah:ToggleSwitch Content="遥操作" IsOn="{Binding IsGamepadEnabled}" Command="{Binding ToggleTeleopCommand}"/>
         </StackPanel>
         <!-- 主控制面板 -->
         <StackPanel Grid.Row="2" Grid.Column="0">
