+++ D5RobotWPF/ViewModel/MainViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
+using CommunityToolkit.Mvvm.Input;
 using CommunityToolkit.Mvvm.Messaging.Messages;
 using D5R;
 using DOF5RobotControl_GUI.Model;
@@ -8,15 +9,6 @@ using System.Windows.Threading;
 
 namespace DOF5RobotControl_GUI.ViewModel
 {
-    public struct JogParams
-    {
-        public JogParams() { }
-
-        public JointSelect Joint { get; set; }
-        public bool IsPositive { get; set; }
-        public bool IsJogTaskSpace { get; set; } = false;
-    };
-
     internal class BottomImgMessage : RequestMessage<CamFrame> { }
 
     public partial class MainViewModel : ObservableObject
@@ -33,6 +25,7 @@ namespace DOF5RobotControl_GUI.ViewModel
         public static readonly Joints AssemblePos2 = new(9000, 14000000, -12000000, 5000000, 0); // 90, 0, 0, 0, 0 -> 90, 14, -12, 5, 0 
         public static readonly Joints AssemblePos3 = new(0, -2500000, 4000000, 7000000, 0); // 0, -2.5, 4, 7, 0
 
+        /***** 依赖服务应用 *****/
         private readonly IRobotControlService _robotControlService;
         private readonly IPopUpService _popUpService;
         private readonly ICameraControlService _cameraCtrlService;
@@ -45,85 +38,6 @@ namespace DOF5RobotControl_GUI.ViewModel
         public Dispatcher Dispatcher { get; private set; }
 
         readonly List<CancellationTokenSource> cancelSourceList = []; // 存储所有的取消源，在 stop 时统一取消
-
-        /***** 机器人系统相关 *****/
-        [ObservableProperty]
-        private bool _systemConnected = false;
-
-        [ObservableProperty]
-        [NotifyPropertyChangedFor(nameof(TargetPx))]
-        [NotifyPropertyChangedFor(nameof(TargetPy))]
-        [NotifyPropertyChangedFor(nameof(TargetPz))]
-        private RoboticState _targetState = new(0, 0, 0, 0, 0);
-
-        [ObservableProperty]
-        private RoboticState _currentState = new(0, 0, 0, 0, 0);
-
-        [ObservableProperty]
-        [NotifyPropertyChangedFor(nameof(TargetPx))]
-        [NotifyPropertyChangedFor(nameof(TargetPy))]
-        [NotifyPropertyChangedFor(nameof(TargetPz))]
-        private bool _isPoseRelative = true;
-
-        public double TargetPx
-        {
-            get
-            {
-                if (IsPoseRelative)
-                    return TargetState.TaskSpace.Px - 72.90;
-                else
-                    return TargetState.TaskSpace.Px;
-            }
-            set
-            {
-                if (IsPoseRelative)
-                    TargetState.TaskSpace.Px = value + 72.90;
-                else
-                    TargetState.TaskSpace.Px = value;
-
-                OnPropertyChanged();
-            }
-        }
-
-        public double TargetPy
-        {
-            get
-            {
-                if (IsPoseRelative)
-                    return TargetState.TaskSpace.Py + 88.75;
-                else
-                    return TargetState.TaskSpace.Py;
-            }
-            set
-            {
-                if (IsPoseRelative)
-                    TargetState.TaskSpace.Py = value - 88.75;
-                else
-                    TargetState.TaskSpace.Py = value;
-
-                OnPropertyChanged();
-            }
-        }
-
-        public double TargetPz
-        {
-            get
-            {
-                if (IsPoseRelative)
-                    return TargetState.TaskSpace.Pz + 88.46;
-                else
-                    return TargetState.TaskSpace.Pz;
-            }
-            set
-            {
-                if (IsPoseRelative)
-                    TargetState.TaskSpace.Pz = value - 88.46;
-                else
-                    TargetState.TaskSpace.Pz = value;
-
-                OnPropertyChanged();
-            }
-        }
 
         [ObservableProperty]
         private bool _camMotorIsConnected = false;
@@ -151,6 +65,9 @@ namespace DOF5RobotControl_GUI.ViewModel
 
         ~MainViewModel()
         {
+            DisconnectSystem();
+            OpcDisconnect();
+
             insertCancelSource?.Cancel();
             attachCancelSource?.Cancel();
 
@@ -160,30 +77,56 @@ namespace DOF5RobotControl_GUI.ViewModel
             }
         }
 
-        /***** OPC 相关代码 *****/
-
-        public void OpcMapMethod(int method)
+        /// <summary>
+        /// 打开配置窗口
+        /// </summary>
+        [RelayCommand]
+        private static void OpenConfigWindow()
         {
-            // 创建一个 EventArgs 实例来传递给 ButtonClicked 方法
-            switch (method)
+            ConfigWindow window = new()
             {
-                case 1: break;
-                case 2: ToggleConnectCommand.Execute(null); break;
-                case 3: SetTargetJointsCommand.Execute(ZeroPos); break;
-                case 4: SetTargetJointsCommand.Execute(IdlePos); break;
-                case 5: SetTargetJointsCommand.Execute(PreChangeJawPos); break;
-                case 6: SetTargetJointsCommand.Execute(ChangeJawPos); break;
-                case 7: SetTargetJointsCommand.Execute(AssemblePos1); break;
-                case 8: SetTargetJointsCommand.Execute(AssemblePos2); break;
-                case 9: SetTargetJointsCommand.Execute(AssemblePos3); break;
-                case 10: SetTargetJointsCommand.Execute(PreFetchRingPos); break;
-                case 11: SetTargetJointsCommand.Execute(FetchRingPos); break;
-                case 12: RobotRunCommand.Execute(null); break;
-                case 13: RobotStopCommand.Execute(null); break;
-                case 14: RobotSetZeroCommand.Execute(null); break;
+                Owner = Application.Current.MainWindow
+            };
+            window.Show();
+        }
+
+        /// <summary>
+        /// 连接/断开系统
+        /// </summary>
+        [RelayCommand]
+        private void ToggleConnect()
+        {
+            // 若系统未连接，则建立连接
+            if (!SystemConnected)  
+                ConnectSystem();
+            else  // 否则断开连接
+                DisconnectSystem();
+        }
+
+        private void ConnectSystem()
+        {
+            try
+            {
+                _robotControlService.Connect(Properties.Settings.Default.RmdPort);
+                _cameraCtrlService.ConnectCamMotor(Properties.Settings.Default.CamMotorPort);
+                StartUpdateState();
+                SystemConnected = true;
+            }
+            catch (InvalidOperationException ex)
+            {
+                _robotControlService.Disconnect();
+                _cameraCtrlService.DisconnectCamMotor();
+                _popUpService.Show(ex.Message);
+                throw;
             }
         }
 
-        /***** OPC 相关代码结束 *****/
+        private void DisconnectSystem()
+        {
+            StopUpdateState();
+            _robotControlService.Disconnect();
+            _cameraCtrlService.DisconnectCamMotor();
+            SystemConnected = false;
+        }
     }
 }
